ETL-Pipeline-Anas/submission-etl-anas/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes
-e .
beautifulsoup4==4.14.2
    # via submission-etl-anas
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
coverage==7.11.0
    # via pytest-cov
google-auth==2.42.1
    # via
    #   google-auth-oauthlib
    #   gspread
    #   submission-etl-anas
google-auth-oauthlib==1.2.2
    # via gspread
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
gspread==6.2.1
    # via submission-etl-anas
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
numpy==2.3.4
    # via pandas
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via pytest
pandas==2.3.3
    # via submission-etl-anas
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.11
    # via submission-etl-anas
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via pytest-cov
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   submission-etl-anas
python-dotenv==1.2.1
    # via submission-etl-anas
pytz==2025.2
    # via pandas
requests==2.32.5
    # via
    #   requests-oauthlib
    #   submission-etl-anas
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via submission-etl-anas
tabulate==0.9.0
    # via submission-etl-anas
tomli==2.3.0 ; python_full_version <= '3.11'
    # via coverage
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
